enhancement(vrl): add generic `encode_key_value`
